implemented streaming for songs and added song duration calculation, …

+++ backend/src/main/java/com/codecool/tavirutyutyu/zsomlexd/controller/dto/SongDTO.java
@@ -1,4 +1,4 @@
 package com.codecool.tavirutyutyu.zsomlexd.controller.dto;
 
-public record SongDTO(String title, String audioBase64, int length, int numberOfLikes, int reShares) {
+public record SongDTO(String title, String audioBase64, double length, int numberOfLikes, int reShares) {
 }
